requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djoser==2.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
gunicorn==23.0.0
pillow==11.0.0
